Test: Split snippet matchers into framework files with exhaustive dispatch
Each framework's represented-names matcher now lives in its own
snippets-<framework>.ts, with the shared root-element and attribute
scanning in parse-element.ts. The dispatch in snippets.ts is an
exhaustive switch over the Framework union: adding a framework fails
compilation at the switch until its matcher exists (verified: extending
the union to 'svelte' fails vue-tsc with TS2322 at that line).
